Fix casing of logoutURL shibboleth property
`Logouturl` does not match cases with `logoutURL`, which causes the attribute request to work with `useHeaders` (because everything is lowercased anyways and headers are case-insensitive). However, when using environment variables, the attribute is undefined, resulting in a logout URL that only executes a local logout and not a full one.